move docker cli func to separate file
